Refactor building dataframe and set specific cluster order
The way I was building the dataframe to be fed to boxplot was kind of
ugly. I just included some of the dictionaries needed to map old color
cluster names to newer cell type names and maps to the specific colors
to use in the notebook instead of just reading external files.